+++ src/features/docs/index.ts
@@ -1,7 +1,2 @@
-export { useDockerCategories } from "./model";
-export {
-  CommentList,
-  CommentWrite,
-  SelectedStepThread,
-  ThreadCard,
-} from "./ui";
+export { useGetDocsList } from "./model/useGetDocsList";
+export { DocsList } from "./ui/DocsList";
