--- src/features/docs/model/index.ts
@@ -1,2 +1,0 @@
-export { useDockerCategories } from "./useDockerCategories";
-export type { Category } from "./types";
