--- src/features/docs/model/types.ts
@@ -1,6 +1,0 @@
-export interface Category {
-  id: string;
-  name: string;
-  parentId: string | null;
-  groupId: string | null;
-}
