+++ src/features/docs/model/useGetDocsList.ts
@@ -3,8 +3,8 @@ import { categoryKeys } from "@/entities/category";
 import type { DocsCategory } from "@/shared/types";
 import { transformCategoriesToDocs } from "../lib/transformCategoriesToDocs";
 
-export function useGetDocsList(): DocsCategory[] {
+export const useGetDocsList = (): DocsCategory[] => {
   const { data: categories = [] } = useQuery(categoryKeys.list());
 
   return transformCategoriesToDocs(categories);
-}
+};
