+++ src/features/docs/ui/DocsList.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "@tanstack/react-router";
 import type { DocsCategory } from "@/shared/types";
 
-export function DocsSectionList({ docs }: { docs: DocsCategory[] }) {
+export const DocsList = ({ docs }: { docs: DocsCategory[] }) => {
   const navigate = useNavigate();
 
   const goToDetail = (category: string, chapterId: string) => {
@@ -17,9 +17,7 @@ export function DocsSectionList({ docs }: { docs: DocsCategory[] }) {
         <div key={section.id} className="mb-12">
           {section.chapters.map((chapter) => (
             <div key={chapter.id} className="mb-6">
-              <h3 className="text-xl font-semibold mb-2">
-                {chapter.title}
-              </h3>
+              <h3 className="text-xl font-semibold mb-2">{chapter.title}</h3>
               <ul className="space-y-1 ml-4">
                 {chapter.steps.map((step) => (
                   <li
@@ -37,4 +35,4 @@ export function DocsSectionList({ docs }: { docs: DocsCategory[] }) {
       ))}
     </>
   );
-}
+};
