--- src/features/docs/ui/index.ts
@@ -1,4 +1,0 @@
-export { CommentList } from "./CommentList";
-export { CommentWrite } from "./CommentWrite";
-export { SelectedStepThread } from "./SelectedStepThread";
-export { ThreadCard } from "./ThreadCard";
